Add ability to compress ocean and ice products (#3111)
This PR adds the ability to optionally compress ocean and ice products
for the GEFS reforecast. Compression can be set using a new
`config.base` variable called `DO_OCNICE_COMPRESS`. For the reforecast,
the ice history data and the 0p25 ocean post-processed data are gzipped
at the end of the ice and ocean prod tasks, respectively. The archive
task was also modified to allow either gzipped or non-gzipped ocean and
ice data to be sent to HPSS, which is controlled by
`DO_OCNICE_COMPRESS`.

Furthermore, an option was added to skip the ocean and ice
post-processing within the ocean and ice prod jobs, respectively. This
can be controlled by `DO_ICE_INTERP` and `DO_OCN_INTERP` in
`config.base`. In order to meet the reforecast requirements,
`DO_ICE_INTERP` is set to `NO` and and `DO_OCN_INTERP` is set to `YES`
by default.

This PR also accomplishes the following:

- [x] The default WCOSS2 `STMP` and `PTMP` were modified to
`/lfs/h3/emc/gefstemp/` because this is the space where the reforecast
will be performed.
- [x] The wavepostpnt task has been removed. 
- [x] If a group account is used, divide the stmp and ptmp directories
by `SUDO_USER`.
- [x] The clean_up job has been modified for the reforecast.