Update param files for GEFS (#3677)
The variables within the GEFS atmos param files have been rearranged so
that the "a" group and "b" group param files are consistent with the
corresponding GEFSv12 param files for both the f000 and fFFF lists. Most
of the variables that have been rearranged are those that currently
exist in both GEFSv12 and will exist in GEFSv13. The same modifications
have been made for the 0p25 and 0p50 param files.

This PR will remain a draft until further verification confirms that the
parameter lists contain the correct variables in the "a" and "b" groups.

Resolves #3652